requirements.txt
# ElevenLabs SDK with audio support
elevenlabs[pyaudio]>=1.0.0

# Discord.py with voice support
discord.py[voice]>=2.3.0

# Voice receive extension for Discord
discord-ext-voice-recv>=0.5.0

# Audio processing
pydub>=0.25.1
numpy>=1.24.0

# Async utilities
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# WebSocket for custom ElevenLabs connection
websockets>=12.0

# FFmpeg Python bindings (optional, for audio conversion)
ffmpeg-python>=0.2.0
